feat: allows JSON type overrides in `Decode()` (#9)
With this change, `Decode()` can take an optional list of Go types and their
respective JSON types which it'll use while decoding the provided payload.

**Motivation:**
By default, Go structs are treated as JSON objects by the jsonschema validator.
However, there may be instances when we may want to override the JSON representation
of a type. For example, we may have a custom struct that serves as our own abstraction
around a numeric type. The type itself may be a bit more complex than a simple `int`
or a `float` but, much like Go's built-in types, we may still simply want to transit
its underlying value as a number and deal with the specifics of the abstraction at the
application or API level.